com.unity.animation.cs-jobs-samples@0.5.0-preview # Changelog
All notable changes to this package will be documented in this file.

The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).

## [0.1.0] - 2018-05-07

### This is the first release of *Unity Package Animation C# Jobs Samples*.

First set of examples for the Animation C# Jobs feature. It includes
CutsomMixer, FullBodyIK, and TwoBoneIK.